+++ openbank-notification-service/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.18.0](https://github.com/JiRaska/open-bank-oss/compare/notification-service-v0.17.4...notification-service-v0.18.0) (2026-08-06)
+
+
+### Features
+
+* **campaign,notification:** tell a delivered message from an accepted handoff (ADR-0239) ([#3838](https://github.com/JiRaska/open-bank-oss/issues/3838)) ([789d662](https://github.com/JiRaska/open-bank-oss/commit/789d6626fde13ab48cd72806a07b7c4e6c19ed0e))
+
 ## [0.17.4](https://github.com/JiRaska/open-bank-oss/compare/notification-service-v0.17.3...notification-service-v0.17.4) (2026-08-03)